Started adding MBean support for FmConnection, but commented it out after realizing that it won't compile under 1.4 JDK
FmXmlRequest now throws an exception if you try to do a SELECT for a field that doesn't exist. Previously it just returned null values for that missing field.

Moved several logging statements form FmStatement and FmPreparedStatement into StatementProcessor, where it only needs to be called once.

Added a few new error code descriptions to ErrorCodes.txt

Fixed white space on a few files.

FmRecord now no longer throws exceptions if it can't parse a date / time / timestamp; it prints a stack trace and returns null. This is because there is so much bad data in AMS.