FmConnection.prepareStatement(String, int) now respects the setting of whether or not to return auto-generated keys. This should be used whenever you want to get the result of an INSERT or UPDATE statement. FmStatement.close() now closes the associated ResultSet. This is the correct behavior according to the JavaDocs.
StatementProcessor has been substantially changed with regard to getGeneratedKeys(). This now works with both INSERT and UPDATE statements (used to only work with INSERT), and it no longer only returns just the values that have changed, it now just returns the entire ResultSet that is returned by FileMaker with the new values. This makes it much more efficient, for example, to get the mod count after updating a record, or get the values of calculated fields. This may break backwards compatibility with code that expects to only get the changed values in the result set IF that code uses numeric indexes instead of field names to read the results.