Now works with HSQL for SyncJoinStore. Very fast!!!
SyncServer no longer throws an exception in getStatus(), instead it embeds the text of the exception. This solves the problem of having exceptions thrown which are not in the classpath of the client.

HamsterNode now works with non-text primary keys

SqlServerInfo.java - close ResultSet in a few places where it was being left open.

Engine.java - fixed a bug where records were not being incremented / decremented correctly, which led to cases of records being scanned for deletion when not necessary.

SyncJoin is now an interface - the class was renamed to SyncJoinSerialized