Incremented to 1.145
LoginFrame.java: Fixed a problem with creating the URL to use for HTTP tunneling

MessageToClient.java: Fixed a problem with escaping that was causing a parse error in FileMaker 12 when there were warnings with carriage returns (ie. when the same record is deleted in both nodes and then synced)

StatementProcessor.java: Fixed a problem where layout name was being doubly URLEncoded.