Incremented version to 1.2073
Changes for international date and number formats

FileMakerServerInfo.java:
* Ignore SQL exceptions while doing database validation checks

FileMakerXmlScriptGenerator.java:
* Changed how $isFmp12 var worked to better handle international string formatting when parsing app version
* Changes to how POST/GET is handled in web viewer; seems to be much more reliable for iPhone sync by combining the $response and $response_address assignemnt into a single statement. Also includes a 60 second auto-retry (untested)

FmMetaData.java: insertEmptyRecord() now retries without any embedded values if it gets a 201 when it tries to write row. This happens for serial numbers that have 'required' and 'prohibit modification' both checked

DumpFilter.java: Fixed sequenceNumber to be static so it's shared across all instances. Also now includes the query portion of the URL in the dump.