Improved some error messages
FmXmlRequest.java: Fixed a condition where you could get this error: "FWK005 parse may not be called while parsing". This could be caused when doing an update of multiple records, because sometimes the parsing thread is still running when the next update starts. Now this is fixed by creating a new SAXParser if the previous one is not finished yet.

StatementProcessor.java: URLEncode field names used in sort order